+++ src/vasu-playwright-lib/utils/page-utils.ts
@@ -6,8 +6,14 @@
  */
 
 import { SMALL_TIMEOUT } from '../constants/timeouts';
-import { Page, expect } from '@playwright/test';
-import { SwitchPageOptions } from '../types/optional-parameter-types';
+import { BrowserContext, Page, Response, expect } from '@playwright/test';
+import {
+  GotoOptions,
+  NavigationOptions,
+  SwitchPageOptions,
+  WaitForLoadStateOptions,
+} from '../types/optional-parameter-types';
+import { getDefaultLoadState } from '../constants';
 
 let page: Page;
 
@@ -94,4 +100,117 @@ export async function closePage(winNum?: number): Promise<void> {
   if (noOfWindows > 1) {
     await switchToDefaultPage();
   }
+}
+
+/**
+ * 1. Navigations: This section contains functions for navigating within a web page or between web pages.
+ * These functions include going to a URL, waiting for a page to load, reloading a page, and going back to a previous page.
+ */
+
+/**
+ * Navigates to the specified URL.
+ * @param {string} path - The URL to navigate to.
+ * @param {GotoOptions} options - The navigation options.
+ * @returns {Promise<null | Response>} - The navigation response or null if no response.
+ */
+export async function gotoURL(
+  path: string,
+  options: GotoOptions = { waitUntil: getDefaultLoadState() },
+): Promise<null | Response> {
+  return await getPage().goto(path, options);
+}
+
+/**
+ * Returns the URL of the page.
+ * @param {NavigationOptions} [options] - Optional navigation options.
+ * @returns {Promise<string>} - The URL of the page.
+ */
+export async function getURL(options: NavigationOptions = { waitUntil: 'load' }): Promise<string> {
+  try {
+    await waitForPageLoadState(options);
+    return getPage().url();
+  } catch (error) {
+    console.log(`getURL- ${error instanceof Error ? error.message : String(error)}`);
+    return '';
+  }
+}
+
+/**
+ * Waits for a specific page load state.
+ * @param {NavigationOptions} options - The navigation options.
+ */
+export async function waitForPageLoadState(options?: NavigationOptions): Promise<void> {
+  let waitUntil: WaitForLoadStateOptions = getDefaultLoadState();
+
+  if (options?.waitUntil && options.waitUntil !== 'commit') {
+    waitUntil = options.waitUntil;
+  }
+
+  await getPage().waitForLoadState(waitUntil);
+}
+
+/**
+ * Reloads the current page.
+ * @param {NavigationOptions} options - The navigation options.
+ */
+export async function reloadPage(options?: NavigationOptions): Promise<void> {
+  await Promise.all([getPage().reload(options), getPage().waitForEvent('framenavigated')]);
+  await waitForPageLoadState(options);
+}
+
+/**
+ * Navigates back to the previous page.
+ * @param {NavigationOptions} options - The navigation options.
+ */
+export async function goBack(options?: NavigationOptions): Promise<void> {
+  await Promise.all([getPage().goBack(options), getPage().waitForEvent('framenavigated')]);
+  await waitForPageLoadState(options);
+}
+
+/**
+ * Waits for a specified amount of time.
+ * @param {number} ms - The amount of time to wait in milliseconds.
+ */
+export async function wait(ms: number): Promise<void> {
+  // eslint-disable-next-line playwright/no-wait-for-timeout
+  await getPage().waitForTimeout(ms);
+}
+
+/**
+ * Retrieves the size of the browser window. This function uses the `evaluate` method to execute code in the context of the page,
+ * allowing it to access the `window` object and retrieve the current window dimensions.
+ * @returns A promise that resolves to an object containing the width and height of the window.
+ */
+export async function getWindowSize(): Promise<{ width: number; height: number }> {
+  return await getPage().evaluate(() => {
+    return {
+      width: window.innerWidth,
+      height: window.innerHeight,
+    };
+  });
+}
+
+/**
+ * Saves the storage state of the current page.
+ *
+ * This function captures the storage state of the page, which includes cookies,
+ * local storage, and session storage. The state can be saved to a file if a path is provided.
+ *
+ * @param {string} [path] - The optional file path where the storage state will be saved.
+ * If not provided, the state will only be returned but not saved to a file.
+ *
+ * @returns {Promise<ReturnType<BrowserContext['storageState']>>} - A promise that resolves to the storage state.
+ *
+ * @example
+ *
+ * // Save storage state to a file
+ * saveStorageState('./state.json');
+ *
+ * // Get storage state without saving to a file
+ * const state = await saveStorageState();
+ *
+ * @see {@link https://playwright.dev/docs/api/class-browsercontext#browser-context-storage-state | Playwright BrowserContext.storageState}
+ */
+export async function saveStorageState(path?: string): Promise<ReturnType<BrowserContext['storageState']>> {
+  return await getPage().context().storageState({ path: path });
 }